requirements.txt
django>=4.2,<4.3
gunicorn
psycopg2-binary
python-dotenv
django-unfold
